Initial Commit
This algorithm considers user-based collaborative filtering. However, users in the same city are considered to compute the similarity values among users. Users are identified by user_id, items by location_id of hotels and the ratings are the hotel cluster id. Currently the algorithm is slow and does not finish on my local machine! Perhaps it needs to be reworked to improve the running time. The dataset is too big, and I am wondering if collaborative filtering can be a good solution for such a problem with big data.